app-demo/requirements.txt
astunparse==1.6.3
click==7.1.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
PyPika==0.48.9
python-dotenv==0.20.0
six==1.16.0
Werkzeug==1.0.1
